+++ source/nvidia/vfx/nvidia-vfx.cpp
@@ -86,7 +86,8 @@ streamfx::nvidia::vfx::vfx::vfx()
 		env_size = GetEnvironmentVariableW(ST_ENV_NVIDIA_VIDEO_EFFECTS_SDK_PATH, nullptr, 0);
 		if (env_size > 0) {
 			buffer.resize(static_cast<size_t>(env_size) + 1);
-			env_size = GetEnvironmentVariableW(ST_ENV_NVIDIA_VIDEO_EFFECTS_SDK_PATH, buffer.data(), buffer.size());
+			env_size = GetEnvironmentVariableW(ST_ENV_NVIDIA_VIDEO_EFFECTS_SDK_PATH, buffer.data(),
+											   static_cast<DWORD>(buffer.size()));
 			sdk_path = std::wstring(buffer.data(), buffer.size());
 		} else {
 			PWSTR   str = nullptr;
